refactor: rename wallet folder
