--- src/pages/WalletForm/index.ts
@@ -1,2 +1,0 @@
-export { type TWalletDataForm } from './types'
-export { default as WalletForm } from './WalletForm'
