+++ src/pages/wallet/components/LoadingView.tsx
