+++ src/pages/wallet/components/SettingsCard.tsx
@@ -15,7 +15,7 @@ import { dateErrorCodes, FORM_TYPE_ENUM } from '~/core/constants/form'
 import { getCurrencySymbol } from '~/core/formats/intlFormatNumber'
 import { CurrencyEnum, GetCustomerInfosForWalletFormQuery } from '~/generated/graphql'
 import { useInternationalization } from '~/hooks/core/useInternationalization'
-import { TWalletDataForm } from '~/pages/WalletForm/types'
+import { TWalletDataForm } from '~/pages/wallet/types'
 import { tw } from '~/styles/utils'
 
 interface SettingsCardProps {
