+++ src/pages/wallet/index.ts
@@ -1,0 +1,2 @@
+export { default as CreateWallet } from './CreateWallet'
+export { type TWalletDataForm } from './types'
